requirements.txt
torch
torchvision
Pillow
PyYAML
tqdm
numpy
typing_extensions